LayoutTests/fast/hidpi/image-srcset-invalid-descriptor.html

<html>
<head>
<script>
    window.targetScaleFactor = 1;
</script>
<script src="../../resources/srcset-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<script>
    function runTest() {
        shouldBe('document.getElementById("invalid_with_x").clientWidth', '100');
        shouldBe('document.getElementById("invalid_without_x").clientWidth', '100');
        shouldBe('document.getElementById("future_descriptor").clientWidth', '100');
        shouldBe('document.getElementById("future_descriptor_without_x").clientWidth', '100');
        shouldBe('document.getElementById("x_with_w_ignored").clientWidth', '200');
        shouldBe('document.getElementById("x_with_h_ignored").clientWidth', '200');
    }
</script>
</head>

<body id="body">
    <div>This test passes if the image below is a 100px blue square when the deviceScaleFactor is 1.</div>
    <div> It tests that even though the 1x resource contains many invalid descriptors,
        only the invalid descriptors are ignored (according to the spec).</div>
    <img id="invalid_with_x" src="" srcset="resources/blue-100-px-square.png 43q 1x dfs 3dd, resources/green-400-px-square.png 2x">
    <img id="invalid_without_x" src="" srcset="resources/blue-100-px-square.png 300q, resources/green-400-px-square.png 2x">
    <img id="future_descriptor" srcset=",resources/blue-100-px-square.png 1x future-descriptor(3x, 4h, whatever), resources/green-400-px-square.png 2x">
    <img id="future_descriptor_without_x" srcset=",resources/blue-100-px-square.png future-descriptor(3x, 4h, whatever), resources/green-400-px-square.png 2x">
    <img id="x_with_w_ignored" srcset=",resources/blue-100-px-square.png 1x 100w, resources/green-400-px-square.png 2x">
    <img id="x_with_h_ignored" srcset=",resources/blue-100-px-square.png 1x 100h, resources/green-400-px-square.png 2x">
</body>
</html>
